First Build with Working Neural Network
A straightforward simulation of Mexican Train, a popular domino game, using a Neural Network as an opponent. In its current state, it is not playable by the user, but a new Player object could be created to facilitate this. Currently, the Neural Net opponent needs a lot of data to have any competitive edge over its opponents, and further tweaks will need to be made to improve it